+++ R/sem.missing.paths.R
@@ -11,32 +11,8 @@ sem.missing.paths = function(
   # Add progress bar
   if(.progressBar == T & length(basis.set) > 0) pb = txtProgressBar(min = 0, max = length(basis.set), style = 3) else pb = NULL
   
-  # Identify d-sep tests among endogenous variables and reverse if family != "gaussian"
-  names(basis.set) = 1:length(basis.set)
-  
-  # Get sorted adjacency matrix
-  amat = get.sort.dag(get.formula.list(modelList))
-  
-  # Identify intermediate endogenous variables
-  idx = colnames(amat)[which(colSums(amat[which(colSums(amat) == 0), , drop = FALSE]) > 0)]
-  
-  idx = idx[!idx %in% names(which(colSums(amat[which(!colSums(amat) == 0), , drop = FALSE]) > 0))]
-  
-  # Identify variables in the basis set where intermediate endogenous variables are the response
-  if(any(sapply(modelList[sapply(modelList, function(x) all.vars(formula(x))[1] %in% idx)], function(x) 
-    
-    if(any(class(x) %in% c("glm", "negbin", "glmmPQL", "glmerMod"))) FALSE else class(x) != "gaussian"))) {
-    
-    # Add flag
-    rev = TRUE
-    
-    basis.set = append(basis.set, lapply(which(sapply(basis.set, function(i) i[2] %in% idx)), function(i) 
-      
-      c(basis.set[[i]][2], basis.set[[i]][1], basis.set[[i]][-(1:2)])
-      
-    ) )
-    
-  }
+  # Reverse intermediate endogenous variables fitted to non-normal distributions
+  basis.set = endogenous.reverse(basis.set)
   
   # Perform d-sep tests
   if(length(basis.set) > 0) pvalues.df = do.call(rbind, lapply(1:length(basis.set), function(i) {